requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.8
backcall==0.2.0
black==21.4b2
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
detectron2==0.6+cu101
entrypoints==0.4
executing==1.0.0
fonttools==4.37.1
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
hydra-core==1.2.0
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
iopath==0.1.9
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.5
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
ocrd-fork-pylsd==0.0.4
omegaconf==2.2.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
portalocker==2.5.1
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
stack-data==0.5.0
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tifffile==2022.8.12
toml==0.10.2
tomli==2.0.1
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yacs==0.1.8
zipp==3.8.1
